+++ ldap4j-java/src/main/java/hu/gds/ldap4j/ldap/Syntax.java
@@ -1,0 +1,10 @@
+package hu.gds.ldap4j.ldap;
+
+import org.jetbrains.annotations.NotNull;
+
+public class Syntax {
+    public static final @NotNull String INTEGER_SYNTAX="1.3.6.1.4.1.1466.115.121.1.27";
+
+    private Syntax() {
+    }
+}